# krcg/templates/openapi.yaml
openapi: "3.0.3"
info:
  title: "KRCG"
  version: "{{version}}"
  description: "**V:tES** cards and decks based on VEKN resources, including the [TWDA](http://www.vekn.fr/decks/twd.htm)."
  contact:
    name: "KRCG"
    url: https://github.com/lionel-panhaleux/krcg
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
...
components:
  schemas:
...
    card:
      type: object
      required: ["Id", "Name", "Type", "Card Text", "Set", "Artist"]
      properties:
        Id:
          description: Card ID
          type: string
          minLength: 1
        Name:
          description: Card name
          type: string
          minLength: 1
        Image:
          description: Card image URL
          type: string
          format: url
          minLength: 1
        Aka:
          description: Alternative card name
          type: string
          minLength: 0
        Type:
          description: Card types
          type: array
          minLength: 1
          items:
            description: Type
            type: string
            minLength: 1
        Card Text:
          description: Card text
          type: string
          minLength: 1
        Set:
          description: "Sets the card has been edited in"
          type: array
          minLength: 1
          items:
            description: "[Set]:[Rarity]"
            type: string
            minLength: 1
        Clan:
          description: Clan (crypt cards) or clan requirement (library cards)
          type: array
          minLength: 0
          items:
            description: Clan name
            type: string
            minLength: 1
        Artist:
          description: Card artists
          type: string
          minLength: 1
        Capacity:
          description: Capacity (crypt card) or capacity bonus (library card)
          type: string
          minLength: 0
        Discipline:
          description: Disciplines requirements (library cards only)
          type: array
          minLength: 0
          items:
            description: Discipline name
            type: string
            minLength: 1
        Banned:
          description: Year of banishment
          type: string
        Pool Cost:
          description: Pool cost (library cards only)
          type: string
          minLength: 0
        Blood Cost:
          description: Blood cost (library cards only)
          type: string
          minLength: 0
        Conviction Cost:
          description: Conviction cost (library cards only)
          type: string
          minLength: 0
        Burn Option:
          description: Burn option (library cards only)
          type: boolean
        Flavor Text:
          description: Flavor text (library cards only)
          type: string
          minLength: 0
        Requirement:
          description: Requirement (library cards only)
          type: string
          minLength: 0
        Draft:
          description: Draft text (library card only)
          type: string
          minLength: 0
        Group:
          description: Group (crypt cards only) - can be "ANY"
          type: string
          minLength: 0
        Disciplines:
          description: Disciplines (crypt cards only)
          type: array
          minLength: 0
          items:
            description: Discipline trigram
            type: string
            minLength: 1
        Title:
          description: Title (crypt cards only)
          type: string
          minLength: 0
        Rulings:
          description: Rulings
          type: array
          minLength: 0
          items:
            description: Ruling text
            type: string
            minLength: 1
        Rulings Links:
          description: Official rulings URLs
          type: array
          minLength: 0
          items:
            description: Ruling URL
            type: object
            required: ["Reference", "URL"]
            properties:
              Reference:
                description: Ruling reference (without brackets)
                type: string
                minLength: 1
              URL:
                description: Ruling URL
                type: string
                format: url
                minLength: 1
        Translations:
          description: Card translations
          type: object
          additionalProperties:
            type: object
            required: ["Name", "Card Text"]
            properties:
              Name:
                description: Card name
                type: string
                minLength: 1
              Card Text:
                description: Card text
                type: string
                minLength: 1
              Flavor Text:
                description: Flavor text (library cards only)
                type: string
                minLength: 0
      example:
        Id: "100029"
        Name: "Aid from Bats"
        Image: "https://images.krcg.org/aidfrombats.jpg"
        Aka: ""
        Type: ["Combat"]
        Clan: []
        Discipline: ["Animalism"]
        Pool Cost": ""
        Blood Cost": ""
        Conviction Cost: ""
        Burn Option: False
        Card Text: |
          [ani] Strike: 1R damage, with 1 optional maneuver.
          [ANI] As above, with 1 optional press.
        Flavor Text: |
          Hanging upside down like rows of disgusting old rags
          And grinning in their sleep. Bats!
          D.H. Lawrence, "Bat"
        Set: ["Jyhad:C", "VTES:C", "CE:C/PN3", "Anarchs:PG2", "Third:C", "KoT:C", "FB:PN6"]
        Requirement: ["justicar", "inner circle"]
        Banned: ""
        Artist: "Melissa Benson; Eric Lofgren"
        Capacity: ""
        Draft: ""
        Rulings:
          - "[ANI] The press can only be used during the current round. [TOM 19960521]"
        Rulings Links:
          - Reference: "TOM 19960521"
            URL: "https://groups.google.com/d/msg/rec.games.trading-cards.jyhad/4emymfUPwAM/B2SCC7L6kuMJ"
        Translations:
          fr-FR:
            Name: "Aide des chauves-souris"
            Card Text: |
              [ani] Frapper à toute portée : 1 point de dégâts, avec 1 manœuvre optionnelle.
              [ANI] Comme ci-dessus, avec 1 poursuite optionnelle.
            Flavor Text: |
              Pendues tête en bas comme des rangées de guenilles repoussantes
              Et souriant de toutes leurs dents dans leur sommeil. Des chauves-souris !
              D.H. Lawrence, "La Chauve-souris"
          es-ES:
            Name: "Ayuda de murciélagos"
            Card Text: |
              [ani] Ataque: 1 de daño a distancia, con 1 maniobra opcional.
              [ANI] Como antes, con 1 acoso opcional.
            Flavor Text: |
              Colgando boca abajo como hileras de trapos viejos y repugnantes
              Y sonriendo mientras duermen. ¡Murciélagos!
              D.H. Lawrence, "Murciélago"
    deck:
      type: object
      required: ["twda_id", "date", "player", "crypt", "library"]
      properties:
        twda_id: { type: string }
        name: { type: string }
        event: { type: string }
        place: { type: string }
        date: { type: string, format: date }
        player: { type: string }
        author: { type: string }
        players_count: { type: integer }
        tournament_format: { type: string }
        score: { type: string }
        comments: { type: string }
        crypt:
          type: object
          required: ["count", "cards"]
          properties:
            count: { type: integer }
            cards:
              type: array
              minLength: 1
              items: { $ref: "#/components/schemas/card_count" }
        library:
          type: object
          required: ["count", "cards"]
          properties:
            count: { type: integer }
            cards:
              type: array
              minLength: 1
              maxLength: 20
              items:
                type: object
                required: ["type", "count", "cards"]
                properties:
                  type: { type: string }
                  count: { type: integer }
                  cards:
                    type: array
                    minLength: 1
                    maxLength: 90
                    items: { $ref: "#/components/schemas/card_count" }
      example:
        twd_id: "2014incbi"
        name: "AAA - Alert Attention Alarm"
        event: "Italian NC 2014"
        place: "Bologna, Italy"
        date: "2014-06-01"
        player: "Enrico De Stefano (Embraced)"
        author: null
        players_count: 25
        tournament_format: "3R+F"
        score: null
        comments: "-- Once again, our Enrico De Stefano defend his national title. Congratulations.\n"
        crypt:
          count: 12
          cards:
            [
              { count: 4, name: Anson, id: "200107" },
              { count: 2, name: Alexandra, id: "200049" },
              { count: 2, name: Anneke, id: "200106" },
              { count: 1, name: François Villon, id: "200478" },
              { count: 1, name: Marcellus, id: "200910" },
              { count: 1, name: Tatiana Romanov, id: "201347" },
              { count: 1, name: "Volker, The Puppet Prince", id: "201458" },
            ]
        library:
          count: 90
          cards:
            - type: Master
              count: 30
              cards:
                [
                  { count: 9, name: Ashur Tablets, id: "100106" },
                  { count: 7, name: Liquidation, id: "101112" },
                  { count: 5, name: Villein, id: "102121" },
                  { count: 3, name: Haven Uncovered, id: "100897" },
                  { count: 1, name: Art Museum, id: "100095" },
                  { count: 1, name: "Golconda: Inner Peace", id: "100842" },
                  { count: 1, name: Information Highway, id: "100984" },
                  { count: 1, name: Papillon, id: "101350" },
                  { count: 1, name: "Pentex(TM) Subversion", id: "101384" },
                  { count: 1, name: Wider View, id: "102180" },
                ]
            - type: Action
              count: 3
              cards: [{ count: 2, name: Entrancement, id: "100652" }, { count: 1, name: Heart of the City, id: "100904" }]
            - type: Equipment
              count: 11
              cards:
                [
                  { count: 6, name: ".44 Magnum", id: "100001" },
                  { count: 3, name: Heart of Nizchetus, id: "100903" },
                  { count: 1, name: Bowl of Convergence, id: "100243" },
                  { count: 1, name: Codex of the Edenic Groundskeepers, id: "100374" },
                ]
            - type: Political Action
              count: 4
              cards: [{ count: 3, name: Parity Shift, id: "101353" }, { count: 1, name: Political Stranglehold, id: "101417" }]
            - type: Action Modifier
              count: 2
              cards: [{ count: 2, name: Aire of Elation, id: "100031" }]
            - type: Action Modifier/Combat
              count: 4
              cards: [{ count: 4, name: "Resist Earth's Grasp", id: "101610" }]
            - type: Reaction
              count: 13
              cards: [{ count: 8, name: "Second Tradition: Domain", id: "101706" }, { count: 5, name: Telepathic Misdirection, id: "101949" }]
            - type: Combat
              count: 21
              cards:
                [
                  { count: 4, name: Blur, id: "100227" },
                  { count: 4, name: Concealed Weapon, id: "100392" },
                  { count: 4, name: "Psyche!", id: "101507" },
                  { count: 4, name: Taste of Vitae, id: "101945" },
                  { count: 3, name: Charismatic Aura, id: "100333" },
                  { count: 2, name: Side Strike, id: "101778" },
                ]
            - type: Event
              count: 2
              cards: [{ count: 1, name: "Anthelios, The Red Star", id: "100074" }, { count: 1, name: Scourge of the Enochians, id: "101697" }]
paths:
  /card:
    post:
      tags: [Card]
      summary: Get cards matching given filters.
      description: |
        If multiple filters are specified, only get cards matching all filters.
        If no filter is specified, get all cards.
        Only the "discipline" field is case-sensitive.

        Note that:

        - Crypt / library are included in "type".
        - Sect and titles are included in "trait",
          independant titles are registered as "1 vote" and "2 votes".
        - City titles are registered as a city trait
          with the name of the city, eg. "los angeles".
        - Crypt cards in the "any" group are also registered in all groups.
        - Disciplines are case-sensitive:
          ALL-UPPERCASE to denote the **superior** version
          and all-lowercase to denote the *inferior* version.
        - Only crypt cards register superior disciplines.
        - Multi-disciplines cards are also registered in the "*" discipline.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  description: response format control (return name or full card)
                  type: string
                  enum:
                    - name
                    - full
                type:
                  description: Type of the card
                  type: array
                  items:
                    type: string
                    enum:
                      - action
                      - action modifier
                      - ally
                      - combat
                      - conviction
                      - crypt
                      - equipment
                      - event
                      - imbued
                      - library
                      - master
                      - political action
                      - power
                      - reaction
                      - retainer
                      - vampire
                clan:
                  description: Clan/Creed or required clan/creed (or none)
                  type: array
                  items:
                    type: string
                    enum:
                      - abomination
                      - ahrimane
                      - akunanse
                      - assamite
                      - avenger
                      - baali
                      - blood brother
                      - brujah
                      - brujah antitribu
                      - caitiff
                      - daughter of cacophony
                      - defender
                      - follower of set
                      - gangrel
                      - gangrel antitribu
                      - gargoyle
                      - giovanni
                      - guruhi
                      - harbinger of skulls
                      - innocent
                      - ishtarri
                      - judge
                      - kiasyd
                      - lasombra
                      - malkavian
                      - malkavian antitribu
                      - martyr
                      - nagaraja
                      - none
                      - nosferatu
                      - nosferatu antitribu
                      - osebo
                      - pander
                      - ravnos
                      - redeemer
                      - salubri
                      - salubri antitribu
                      - samedi
                      - toreador
                      - toreador antitribu
                      - tremere
                      - tremere antitribu
                      - true brujah
                      - tzimisce
                      - ventrue
                      - ventrue antitribu
                      - visionary
                group:
                  description: Group (crypt cards only)
                  type: array
                  items:
                    type: string
                    enum:
                      - "1"
                      - "2"
                      - "3"
                      - "4"
                      - "5"
                      - "6"
                      - "any"
                trait:
                  description: Trait, title, city or sect
                  type: array
                  items:
                    type: string
                    enum:
                      - 1 vote
                      - 2 votes
                      - amsterdam
                      - anarch
                      - aragon
                      - archbishop
                      - athens
                      - atlanta
                      - barcelona
                      - baron
                      - berlin
                      - bishop
                      - black hand
                      - boston
                      - brussels
                      - budapest
                      - cairo
                      - camarilla
                      - canberra
                      - cardinal
                      - chicago
                      - chicago
                      - cleveland
                      - columbus
                      - cordoba
                      - corte
                      - dallas
                      - detroit
                      - dublin
                      - fortaleza
                      - frankfurt
                      - geneva
                      - glasgow
                      - guadalajara
                      - guatemala city
                      - houston
                      - imperator
                      - independent
                      - infernal
                      - inner circle
                      - istanbul
                      - justicar
                      - kholo
                      - laibon
                      - liaison
                      - lisbon
                      - london
                      - los angeles
                      - magaji
                      - manila
                      - mexico city
                      - miami
                      - milan
                      - milwaukee
                      - monaco
                      - montreal
                      - new york
                      - paris
                      - perth
                      - philadelphia
                      - pittsburgh
                      - port-au-prince
                      - prague
                      - primogen
                      - prince
                      - priscus
                      - red list
                      - regent
                      - rome
                      - sabbat
                      - san diego
                      - seattle
                      - seraph
                      - slave
                      - stockholm
                      - strasbourg
                      - sydney
                      - toronto
                      - venice
                      - versailles
                      - washington, d.c.
                discipline:
                  description: Discipline
                  type: array
                  items:
                    type: string
                    enum:
                      - "*"
                      - ANIMALISM
                      - AUSPEX
                      - CELERITY
                      - CHIMERSTRY
                      - DAIMOINON
                      - DEMENTATION
                      - DOMINATE
                      - FORTITUDE
                      - MELPOMINEE
                      - MYTHERCERIA
                      - NECROMANCY
                      - OBEAH
                      - OBFUSCATE
                      - OBTENEBRATION
                      - POTENCE
                      - PRESENCE
                      - PROTEAN
                      - QUIETUS
                      - SANGUINUS
                      - SERPENTIS
                      - SPIRITUS
                      - TEMPORIS
                      - THANATOSIS
                      - THAUMATURGY
                      - VALEREN
                      - VICISSITUDE
                      - VISCERATIKA
                      - abombwe
                      - animalism
                      - auspex
                      - celerity
                      - chimerstry
                      - daimoinon
                      - defense
                      - dementation
                      - dominate
                      - flight
                      - fortitude
                      - innocence
                      - judgment
                      - maleficia
                      - martyrdom
                      - melpominee
                      - mytherceria
                      - necromancy
                      - none
                      - obeah
                      - obfuscate
                      - obtenebration
                      - potence
                      - presence
                      - protean
                      - quietus
                      - redemption
                      - sanguinus
                      - serpentis
                      - spiritus
                      - striga
                      - temporis
                      - thanatosis
                      - thaumaturgy
                      - valeren
                      - vengeance
                      - vicissitude
                      - visceratika
                      - vision
                bonus:
                  description: Special bonus provided
                  type: array
                  items:
                    type: string
                    enum:
                      - votes
                      - intercept
                      - bleed
                      - stealth
                text:
                  description: Card text to search for - also matches card names
                  type: string
                lang:
                  description: Additional language to search in (for text search only)
                  type: string
                  format: language tag
              example:
                type: ["political action"]
                trait: ["anarch"]
      responses:
        "200":
          description: Success, returns a JSON array of card names
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    minLength: 1
                    items:
                      description: Card name
                      type: string
                  - type: array
                    minLength: 1
                    items: { $ref: "#/components/schemas/card" }
                example:
                  - Anarch Salon
                  - Eat the Rich
                  - Exclusion Principle
                  - Firebrand
                  - Free States Rant
                  - Patsy
                  - Persona Non Grata
                  - Reckless Agitation
                  - Revolutionary Council
                  - Sweepe
        "404":
          description: Not found
  /card/{name_or_id}:
    get:
      tags: [Card]
      summary: Get card text and rulings
      description: Return the matching card information
      parameters:
        - name: name_or_id
          in: path
          required: true
          schema:
            type: string
            description: "Card name or ID"
      responses:
        "200":
          description: Success, returns the card as JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/card" }
        "404":
          description: Not found
  /complete/{partial_name}:
    get:
      tags: [Card]
      summary: Complete a card name
      description: Return all candidates matching the partial name
      parameters:
        - name: partial_name
          in: path
          required: true
          schema:
            type: string
            description: "Partial card name"
      responses:
        "200":
          description: Success, returns a JSON array of card names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Card name
              example: ["The unnamed", "Unnatural Disaster"]
        "404":
          description: Not found
  /complete_i18n/{partial_name}:
    get:
      tags: [Card]
      summary: Complete a card name with translations
      description: |
        Return all candidates matching the partial name, with translations.
        Uses the language provided in Accept-Language (best match) and
        matches on the original english names as well as translated names.
      parameters:
        - name: partial_name
          in: path
          required: true
          schema:
            type: string
            description: "Partial card name"
        - name: Accept-Language
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success, returns a JSON array of card names translations (dict)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  description: Card name and translation
                  additionalProperties:
                    type: string
              example: [{ "en": "Aid from Bats", "fr-FR": "Aide des chauves-souris" }]
        "404":
          description: Not found
  /submit-ruling/{card_name_or_id}:
    post:
      tags: [Card]
      summary: Submit a new ruling for examination
      parameters:
        - name: card_name_or_id
          in: path
          required: true
          description: "Card name or ID"
          schema:
            type: string
            description: "Card name or ID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  description: "Envisioned text for the ruling"
                  type: string
                link:
                  description: "Link to the official ruling from a Rules Director"
                  type: string
                  format: url
              example:
                text: Cannot be played by vampires who aren't controlled.
                link: https://groups.google.com/d/msg/rec.games.trading-cards.jyhad/80KRDjVFkyg/TBJso9Ra_9QJ
      responses:
        "201":
          description: Successful, request created
        "400":
          description: Malformed body
        "404":
          description: Card not found
  /deck:
    post:
      tags: [Deck]
      summary: Get TWDA decks matching given filters
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                card_names:
                  description: "If absent, all decks are returned"
                  type: array
                  minLength: 0
                  items:
                    description: "Case-insensitive card name"
                    type: string
                date_from:
                  description: "Only consider decks from this date on"
                  type: string
                  format: date
                  default: "2008-01-01"
                date_to:
                  description: "Only consider decks until this date"
                  type: string
                  format: date
                  default: null
                players:
                  description: "Only consider tournaments with at least this number of players"
                  type: integer
                  minimum: 0
                  default: 0
              example:
                cards: ["François Villon"]
                date_from: "1994-01-01"
                players: 25
      responses:
        "200":
          description: Successful, returns a list of decks as a JSON array
          content:
            application/json:
              schema:
                type: array
                minLength: 1
                items: { $ref: "#/components/schemas/deck" }
        "404":
          description: Not found
  /deck/{twda_id}:
    get:
      tags: [Deck]
      summary: Get a deck by its TWDA ID
      description: Returns the deck with the given ID
      parameters:
        - name: twda_id
          in: path
          required: true
          schema:
            type: string
            description: "TWDA ID"
      responses:
        "200":
          description: Success, returns the deck as JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
        "404":
          description: Not Found